moved termination metrics to termination controller
